backend/summarizer/requirements.txt
pika
openai
python-dotenv
groq
psycopg2-binary
